requirements.txt
streamlit>=1.32.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pydicom>=2.4.0
torch>=2.0.0
torchvision>=0.15.0
